% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gmrf_Qfuncs.R
\name{Dar}
\alias{Dar}
\title{Differences defining a cyclic auto regressive process
or equivalently, an AR with toroidal edge effects.}
\usage{
Dar(n, phi)
}
\arguments{
\item{n}{the size of the GMRF}

\item{phi}{the autoregressive parameters}
}
\value{
a Matrix
}
\description{
Details
}
\details{
Description - This function does stuff.
}

