% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gmrf_Qfuncs.R
\name{Dharm}
\alias{Dharm}
\title{Differences defining a cyclic nth order random walk}
\usage{
Dharm(n, wavelength = n, cyclic = FALSE)
}
\arguments{
\item{n}{the size of the GMRF}

\item{wavelength}{the wavelength of the harmonic. default is for this
this to be n, i.e. one cycle.}

\item{cyclic}{logical: should the smoother be cyclic, i.e.
corrected for toroidal edge effects.}
}
\value{
a Matrix
}
\description{
Details
}
\details{
Description - This function does stuff.
}

