% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gmrf_Qfuncs.R
\name{Dnb}
\alias{Dnb}
\title{Differences defining a regional model}
\usage{
Dnb(nb)
}
\arguments{
\item{nb}{the neighbourhood structure of a regional layout}
}
\value{
a Matrix with a column for each region and a row for each connection / edge
}
\description{
Details
}
\details{
Description - This function does stuff.
}

