% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gmrf_Qfuncs.R
\name{Drw}
\alias{Drw}
\title{Differences defining a cyclic nth order random walk}
\usage{
Drw(n, order = 2, cyclic = FALSE)
}
\arguments{
\item{n}{the size of the GMRF.}

\item{order}{the order of the random walk.}

\item{cyclic}{logical: should the differences be cyclic, i.e.
corrected for toroidal edge effects.}
}
\value{
a Matrix
}
\description{
Details
}
\details{
Description - This function does stuff.
}

