% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gmrf_Qfuncs.R
\name{Dseasonal}
\alias{Dseasonal}
\title{Differences defining a seasonal model}
\usage{
Dseasonal(n, m)
}
\arguments{
\item{n}{the size of the GMRF}

\item{m}{the length of the seasonal period}
}
\value{
a Matrix
}
\description{
This model treats the sum over m time points
to be stationary
}
\details{
Description - This function does stuff.
}

