% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gmrf_Qfuncs.R
\name{getFactorsnb}
\alias{getFactorsnb}
\title{Get a constraint matrix for a regional GMRF}
\usage{
getFactorsnb(Q)
}
\arguments{
\item{Q}{a precision matrix for a regional GMRF}
}
\value{
a Matrix with a column for each region and a row for each distinct group
}
\description{
Details. Each group of regions sums to zero.
}
\details{
Description - This function does stuff.
}

