% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gmrf_Qfuncs.R
\name{getQar}
\alias{getQar}
\title{Differences defining a cyclic auto regressive process
or equivalently, an AR with toroidal edge effects.}
\usage{
getQar(n, phi, weights = NULL)
}
\arguments{
\item{n}{the size of the GMRF}

\item{phi}{the autoregressive parameters}

\item{weights}{weights to be applied to the node differences in effect
allowing different variances at each time step.}
}
\value{
a Matrix
}
\description{
Details
}
\details{
Description - This function does stuff.
}

