% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gmrf_Qfuncs.R
\name{getQrw}
\alias{getQrw}
\title{Compute RWn precision matrix}
\usage{
getQrw(n, order = 2, weights = NULL, cyclic = FALSE)
}
\arguments{
\item{n}{the size of the GMRF}

\item{order}{the order of the random walk}

\item{weights}{weights to be applied to the node differences (see details)}

\item{cyclic}{logical: should the smoother be cyclic, i.e.
corrected for toroidal edge effects.}
}
\value{
what does it return
}
\description{
Details
}
\details{
Description - This function does stuff.
}

