% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gmrf_Qfuncs.R
\name{getQrw1}
\alias{getQrw1}
\title{Compute RW1 precision matrix}
\usage{
getQrw1(n, weights = NULL, cyclic = FALSE)
}
\arguments{
\item{n}{the size of the GMRF}

\item{weights}{weights to be applied to the node differences (see details)}

\item{cyclic}{logical: should the smoother be cyclic, i.e.
corrected for toroidal edge effects.}
}
\value{
what does it return
}
\description{
If weights are supplied
}
\details{
Description - This function does stuff.
}

