% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gmrf_Qfuncs.R
\name{getRegionalGMRF}
\alias{getRegionalGMRF}
\title{Differences defining a regional model}
\usage{
getRegionalGMRF(nbmat, weights = NULL)
}
\arguments{
\item{nbmat}{the neighbourhood structure of a regional layout as a matrix}

\item{weights}{weights to be applied to the node differences in effect
allowing different connections between regions to vary differently.
An example of a weight could be the length of the shared border between regions}
}
\value{
a Matrix with a column for each region and a row for each connection / edge
}
\description{
This one is for back compatability
}
\details{
Description - This function does stuff.
}

