% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Package.R
\docType{package}
\name{gmrf}
\alias{gmrf}
\alias{gmrf-package}
\title{Lightweight add-on for mgcv to allow slighly easier use of GMRF smoothing models}
\description{
Lightweight add-on for mgcv to allow slighly easier use of GMRF smoothing models
}

