% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gmrf_sim.R
\name{simeQ}
\alias{simeQ}
\title{Simulate from an inhomogenous GMRF}
\usage{
simeQ(eQ, tol = 1e-09, rank = NULL, k = 1)
}
\arguments{
\item{eQ}{an eigen decomposition of a symmetric positive semi-definate matrix corresponding
to the precision matrix of an inhomogenous GMRF}

\item{tol}{tolerance (relative to largest eigen value) for numerical lack
of positive-definiteness in `Q'.}

\item{rank}{the rank of the precision matrix. If this is not supplied it
it is estimated by comparing the eigen values to tol * largest eigenvalue.}

\item{k}{optional scaling of the precision matrix.  This can be used to save
recomputing the eigen decomposition for different smoothing parameters.}
}
\value{
a single draw from with the appropriate conditional covariance structure
}
\description{
Required the eigen decomposition of a precision matrix
}
\details{
Note if rank is suplied and is less than that true rank, the simulation
will be from a reduced rank GMRF.
}
\examples{
## create a rw2 GMRF precision matrix and simulate
Q <- getQrw(100, order = 2)
eQ <- eigen(Q)
x <- simeQ(eQ, k = exp(5))
plot(x, type = "l")
}

