% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gmrf_sim.R
\name{simQ}
\alias{simQ}
\title{Simulate from an inhomogenous GMRF}
\usage{
simQ(Q, tol = 1e-09, rank = NULL)
}
\arguments{
\item{Q}{a symmetric positive semi-definate matrix corresponding
to the precision matrix of an inhomogenous GMRF}

\item{tol}{tolerance (relative to largest eigen value) for numerical lack
of positive-definiteness in `Q'.}

\item{rank}{the rank of the precision matrix. If this is not supplied it
it is estimated by comparing the eigen values to tol * largest eigenvalue}
}
\value{
a single draw from with the appropriate covariance structure
}
\description{
Details
}
\details{
This function does stuff.
}
\examples{
## create a rw2 GMRF precision matrix and simulate
Q <- getQrw(100, order = 2)
x <- simQ(Q)
plot(x)
##
## simulate an AR1 GMRF with toroidal edge correction
Q <- getQar(100, phi = 0.8)
x <- simQ(exp(5)*Q)
plot(x, type = "l")
}

