\name{summaryGmrfFitNugget}
\alias{summaryGmrfFitNugget}
\alias{summaryGmrfFit}
\docType{data}
\title{
Summarize a GMRF fit
}
\description{
Obtain MLE and profile likelihoods from a GMRF model fit.
}
\usage{
summaryGmrfFitNugget(applyResult,MoreArgs)
summaryGmrfFit(applyResult,MoreArgs)
}
\arguments{
\item{applyResult}{A matrix or array from the \code{mapply}
 function run on loglikGmrf or loglikGmrfNugget}
\item{MoreArgs}{A list of additional arguments which \code{mapply} was given}
}


\examples{

}
\keyword{datasets}
