% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gmrf_smooth_spec.R
\name{Predict.matrix.gmrf.smooth}
\alias{Predict.matrix.gmrf.smooth}
\title{Estimate capture probabilites from electrofishing data}
\usage{
\method{Predict.matrix}{gmrf.smooth}(object, data)
}
\arguments{
\item{object}{an object of class `"gmrf.smooth"' produced by the
`smooth.construct' method.}

\item{data}{a list containing just the data (including any `by' variable)
required by this term, with names corresponding to
`object$term' (and `object$by'). The `by' variable is the
last element.}
}
\value{
A design matrix
}
\description{
This function uses the marginal likelihood of capture probabilities
to estimate model parameters
}

