% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gmrf_smooth_spec.R
\name{smooth.construct.gmrf.smooth.spec}
\alias{smooth.construct.gmrf.smooth.spec}
\title{Construct a smoother using a GMRF smoothing prior}
\usage{
\method{smooth.construct}{gmrf.smooth.spec}(object, data, knots)
}
\arguments{
\item{object}{a smooth specification object, usually generated by a term
    `s(...,bs="gmrf", penalty = list(Q = ...))'. `x' is a factor
variable giving labels for the nodes in the GMRF graph, and the `xt'
argument is obligatory: see details.}

\item{data}{a list containing just the data (including any `by' variable)
required by this term, with names corresponding to
`object$term' (and `object$by'). The `by' variable is the
last element.}

\item{knots}{not used.}
}
\value{
An object of class `"gmrf.smooth"'.
}
\description{
This function is used internally in mgcv when fitting a smoother
of type gmrf.  More details to be added.
}
\examples{
require(mgcv)
n <- 100
set.seed(2)
x <- runif(n)
y <- x + x^2*.2 + rnorm(n) *.1
## is smooth significantly different from straight line?
summary(gam(y~s(x,m=c(2,0))+x,method="REML")) ## not quite
## is smooth significatly different from zero?
summary(gam(y~s(x),method="REML")) ## yes!
## see ?gam
}

