% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gmrf_smooth_spec.R
\name{smooth.construct.gmrf.smooth.spec}
\alias{smooth.construct.gmrf.smooth.spec}
\title{Construct a smoother using a GMRF smoothing prior}
\usage{
\method{smooth.construct}{gmrf.smooth.spec}(object, data, knots)
}
\arguments{
\item{object}{a smooth specification object, usually generated by a term
    `s(...,bs="gmrf", penalty = list(Q = ...))'. `x' is a factor
variable giving labels for the nodes in the GMRF graph, and the `xt'
argument is obligatory: see details.}

\item{data}{a list containing just the data (including any `by' variable)
required by this term, with names corresponding to
`object$term' (and `object$by'). The `by' variable is the
last element.}

\item{knots}{not used.}
}
\value{
An object of class `"gmrf.smooth"'.
}
\description{
This function is used internally in mgcv when fitting a smoother
of type gmrf.  More details to be added.
}
\examples{
require(gmrf)
n <- 100
set.seed(2)
Q <- getQrw(100, order = 2)
x <- simQ(exp(1) * Q)
y <- x + rnorm(n) * 3
## set up variables for smoothing
id <- 1:length(x)
rownames(Q) <- colnames(Q) <- id
## fit an RW2 smoother
g1 <- gam(y ~ s(id, bs = "gmrf", xt = list(penalty = Q)), method="REML")
summary(g1)
plot(id, y)
lines(id, x, col = "blue", lwd = 2)
lines(id, fitted(g1), col = "red", lwd = 2)
sefit <- predict(g1, se = TRUE) $ se.fit
lines(id, fitted(g1) + 2*sefit, col = "red", lty = 2)
lines(id, fitted(g1) - 2*sefit, col = "red", lty = 2)
}

